backend/generated_projects/app-1758480941053/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0